cube tank
class tankMove
