+++ D3DX_FrameWork/D3DX_FrameWork/D3DX_FrameWork.vcxproj
@@ -148,6 +148,7 @@
     <ClInclude Include="D3DX_FrameWork.h" />
     <ClInclude Include="Resource.h" />
     <ClInclude Include="stdafx.h" />
+    <ClInclude Include="tankMove.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
   <ItemGroup>
@@ -160,6 +161,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="tankMove.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="D3DX_FrameWork.rc" />
