+++ D3DX_FrameWork/D3DX_FrameWork/D3DX_FrameWork.vcxproj.filters
@@ -36,6 +36,9 @@
     <ClInclude Include="cDeviceManager.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
+    <ClInclude Include="tankMove.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="stdafx.cpp">
@@ -48,6 +51,9 @@
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="cDeviceManager.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="tankMove.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
   </ItemGroup>
